FSPT-596: block a question being deleted if it has dependencies
This adds a final helper to ensure that a question can't be deleted if
any of the other questions in that form depend on it for their
conditional branching.

Becuase the delete method is in a temporary interface file and the route
itself is a developer route there's not much to hook into in terms of
tests here - happy for the reviewer to reason about if this is
sufficient for now.

+++ app/common/data/interfaces/temporary.py
@@ -13,6 +13,7 @@ from uuid import UUID
 
 from sqlalchemy import select, text
 
+from app.common.data.interfaces.collections import DependencyOrderException, depends_on_question
 from app.common.data.models import (
     Collection,
     Form,
@@ -73,6 +74,9 @@ def delete_form(form: Form) -> None:
 
 
 def delete_question(question: Question) -> None:
+    depends_on_this_question = depends_on_question(question)
+    if depends_on_this_question:
+        raise DependencyOrderException("Cannot delete question that is depended on", depends_on_this_question, question)
     db.session.delete(question)
     question.form.questions.reorder()
     db.session.execute(
